refactor(directive): extract aria cluster to directive-aria.ts (DIR-F7)
Move the auto-aria management out of `directive.ts` (~110 LOC) into a
focused `directive-aria.ts` module: `MANAGED_ARIA_ATTRS`, the
`ariaLockKey` / `ariaScopeKey` symbol slots + `AriaCarrier` type,
`mergeAriaLocks` / `setAriaAttr` / `resolveAriaValue` / `applyAria` /
`setupAria` / `teardownAria`. The exported surface is the four lifecycle
helpers (`setupAria`, `teardownAria`, `applyAria`, `mergeAriaLocks`)
+ the `AriaCarrier` type the directive's hooks cast through.

The SSR path (`vRegisterDynamic.getSSRProps`) previously inlined its
own loop over `MANAGED_ARIA_ATTRS` calling `resolveAriaValue`, leaking
two private symbols across the module boundary. Replaced with a new
`getSSRAriaProps(rv, vnode)` helper colocated with the rest of the
cluster — same logic, behind one named entry point, with no
implementation detail crossing the boundary. `directive.ts`'s
`getSSRProps` collapses to two lines.

`DisplayState` drops from `directive.ts`'s import (only the aria
module needs it now). Behavior-neutral: 43 affected ARIA tests
(aria.test, use-register, checkbox, aria-ssr) stay green unchanged.
